GCP/backend/requirements.txt
# Web Framework
Flask==3.0.0
flask-cors
gunicorn
Werkzeug

# Date and Time
future
python-dateutil

# Google Cloud & Firebase
google-cloud-firestore
firebase-admin

# PDF Processing
PyPDF2
pdfminer.six
pdfplumber

# Machine Learning & Text Processing
scikit-learn
numpy
futures>=3.0.5
typing>=3.7.4

# API & Requests
requests
anthropic

# Utils
python-dotenv
typing
